Move GLScalar to its own module.
